rpa-client/requirements.txt
selenium==4.9.1
websocket-client==1.3.1
PyYAML==6.0
